+++ src/player.js
@@ -1,0 +1,3 @@
+const Player = (name, symbol, score = 0) => ({ name, symbol, score });
+
+export default Player;
